+++ src/main.cc
@@ -41,6 +41,9 @@
 
 #if HAVE_EDIT
 #include <editline/readline.h>
+#elif HAVE_READLINE
+#include <readline/readline.h>
+#include <readline/history.h>
 #endif
 
 using namespace ledger;
@@ -137,7 +140,7 @@ int main(int argc, char * argv[], char * envp[])
 
       bool exit_loop = false;
 
-#if HAVE_EDIT
+#if HAVE_EDIT || HAVE_READLINE
       rl_readline_name = const_cast<char *>("Ledger");
       // TODO: rl_attempted_completion_function = ledger_completion;
 
@@ -158,7 +161,7 @@ int main(int argc, char * argv[], char * envp[])
           add_history(expansion);
         }
 
-#else // HAVE_EDIT
+#else // HAVE_EDIT || HAVE_READLINE
 
       while (! std::cin.eof()) {
         std::cout << global_scope->prompt_string();
@@ -167,7 +170,7 @@ int main(int argc, char * argv[], char * envp[])
 
         char * p = skip_ws(line);
 
-#endif // HAVE_EDIT
+#endif // HAVE_EDIT || HAVE_READLINE
 
         check_for_signal();
 
@@ -178,7 +181,7 @@ int main(int argc, char * argv[], char * envp[])
             global_scope->execute_command_wrapper(split_arguments(p), true);
         }
 
-#if HAVE_EDIT
+#if HAVE_EDIT || HAVE_READLINE
         if (expansion)
           std::free(expansion);
         std::free(p);