+++ NEWS.md
@@ -11,6 +11,8 @@
 
 - Fix related reports when using bucket transactions (ledger/ledger#2220)
 
+- Add support to build ledger with readline
+
 ## 3.3.2 (2023-03-30)
 
 - Fix divide by zero (ledger/ledger#777, ledger/ledger#2207)